+++ src/main/registerIpcHandler.ts
@@ -4,9 +4,12 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
+import Store from 'electron-store';
+
 import { registerGetAppDetails } from '../ipc/appDetails';
 import {
     registerDownloadAllAppsJsonFiles,
+    registerDownloadAppIcon,
     registerDownloadReleaseNotes,
     registerGetDownloadableApps,
     registerGetLocalApps,
@@ -17,7 +20,6 @@ import {
 } from '../ipc/apps';
 import { registerGetConfig } from '../ipc/config';
 import { registerCreateDesktopShortcut } from '../ipc/createDesktopShortcut';
-import { registerDownloadToFile } from '../ipc/downloadToFile';
 import {
     registerCancelUpdate,
     registerCheckForUpdate,
@@ -46,6 +48,7 @@ import {
 } from '../ipc/sources';
 import {
     downloadAllAppsJsonFiles,
+    downloadAppIcon,
     downloadReleaseNotes,
     getDownloadableApps,
     getLocalApps,
@@ -57,7 +60,6 @@ import {
 import { getConfig } from './config';
 import createDesktopShortcut from './createDesktopShortcut';
 import { cancelUpdate, checkForUpdate, startUpdate } from './launcherUpdate';
-import { downloadToFile } from './net';
 import { callRegisteredCallback } from './proxyLogins';
 import { requireModule } from './require';
 import {
@@ -73,9 +75,9 @@ import { addSource, getAllSources, removeSource } from './sources';
 import { getAppDetails, openAppWindow, openLauncherWindow } from './windows';
 
 export default () => {
-    registerGetAppDetails(getAppDetails);
+    Store.initRenderer();
 
-    registerDownloadToFile(downloadToFile);
+    registerGetAppDetails(getAppDetails);
 
     registerGetConfig(getConfig);
 
@@ -102,6 +104,7 @@ export default () => {
     registerOpenLauncher(openLauncherWindow);
 
     registerDownloadAllAppsJsonFiles(downloadAllAppsJsonFiles);
+    registerDownloadAppIcon(downloadAppIcon);
     registerDownloadReleaseNotes(downloadReleaseNotes);
     registerGetDownloadableApps(getDownloadableApps);
     registerGetLocalApps(getLocalApps);